Update content script to spawn one balloon #187
